Updated tasks
I have a hard time deciding if I should embrace having commits only
about tasks or not.

On one hand Im tempted to say it should be code only stuff in the repo.
But they I like devops and having that in sync with the code.

So in that vein having the a snapshot of the tasks over time is not that
odd.

So lets try that out, I will probably make more of these type of commits
in the future.