Fixed taskgit newline issue
I want to avoid editing the formating the user has used for now since I
initially thought of this as a textfile first format. If over time the
usage will be using cli and ai tools to generate and manage this file
then I will consider changing the format.

But for now I want to keep the format as is. But if the file does not
end with a newline we need to add one such that the new task [[task]]
does not begin on the same line as the last task as that makes the file
invalid toml.